LPS-69227 Make it better performant via moving out the logic from a model listener
